[Lab 01] 04 - added handling for add spot button + small refactoring of package
