--- app/src/main/java/com/archaeologicalfieldwork/SpotActivity.kt
@@ -1,14 +1,0 @@
-package com.archaeologicalfieldwork
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-import org.jetbrains.anko.AnkoLogger
-import org.jetbrains.anko.info
-
-class SpotActivity : AppCompatActivity(), AnkoLogger {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(savedInstanceState)
-    info("Spot Activity started..")
-    setContentView(R.layout.activity_spot)
-  }
-}
