+++ app/src/main/res/layout/activity_spot.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.archaeologicalfieldwork.SpotActivity">
+    tools:context="com.archaeologicalfieldwork.activities.SpotActivity">
 
   <RelativeLayout
       android:layout_width="match_parent"
@@ -40,7 +40,7 @@
           android:orientation="vertical">
 
         <EditText
-            android:id="@+id/placemarkTitle"
+            android:id="@+id/spotTitle"
             android:layout_width="match_parent"
             android:layout_height="45dp"
             android:layout_margin="8dp"
@@ -54,7 +54,7 @@
             android:textSize="14sp" />
 
         <Button
-            android:id="@+id/btnAdd"
+            android:id="@+id/btnAddSpot"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_margin="16dp"
